pyproject.toml
[project]
name = "mllib"
version = "0.1.0"
description = "MLlib - Machine Learning Library"
requires-python = ">=3.8"
dependencies = [
    "torch",
    "torchvision",
    "pyyaml",
    "pillow",
    "pydantic>=2.10.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
